Use 'ProseMirror' as text editor, for autosaving and collaborative editing
In foundry v10, a new text editor 'ProseMirror' was added [1], and is
set to replace the tinyMCE editor which is being phased out [2].

This new editor looks a bit nicer, and importantly auto-saves the
changes. My players (and me) would frequently accidentally forget to
save, and just close the character sheet (also eventually it seemed to
just time out). This would cause lost content.

It also has the benefit of allowing 2 people to edit the notes at the
same time, which is a nice bonus.

I think its a much nicer experience!

[1]: https://foundryvtt.com/article/v10-text-editor/
[2]: https://foundryvtt.com/api/classes/foundry.applications.ux.TextEditor.html#create